+++ R/tar_call.R
@@ -1,4 +1,4 @@
-#' @title Currently running `targets` function.
+#' @title Identify the called `targets` function.
 #' @export
 #' @family utilities
 #' @description Get the name of the currently running `targets`
@@ -7,21 +7,21 @@
 #'   by [tar_make()], [tar_visnetwork()], etc.).
 #' @return Character of length 1, name of the currently running `targets`
 #'   interface function. For example, suppose you have a call to
-#'   `tar_function()` inside a target or `_targets.R`. Then if you run
-#'   `tar_make()`, `tar_function()` will return `"tar_make"`.
+#'   `tar_call()` inside a target or `_targets.R`. Then if you run
+#'   `tar_make()`, `tar_call()` will return `"tar_make"`.
 #' @examples
 #' if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
 #' tar_dir({ # tar_dir() runs code from a temporary directory.
-#' tar_function() # NULL
+#' tar_call() # NULL
 #' tar_script({
-#'   message("called function: ", tar_function())
-#'   tar_target(x, tar_function())
+#'   message("called function: ", tar_call())
+#'   tar_target(x, tar_call())
 #' })
 #' tar_manifest() # prints "called function: tar_manifest"
 #' tar_make() # prints "called function: tar_make"
 #' tar_read(x) # "tar_make"
 #' })
 #' }
-tar_function <- function() {
+tar_call <- function() {
   tar_runtime$get_fun()
 }